fix: separate cdk stack props from business logic config, fix stack props
This commit fixes an issue where the cdk StackProps were not being
consumed at all.  They were passed to our stack constructor but not
through to the superclass constructor.

To make the constructor arguments more clear, we should not be
extending cdk.StackProps; we should retain separate constructor
args for StackProps vs. the options that are related to our business
logic.  This commit also addresses that issue by removing the
inheritance and splitting out a separate constructor arg for the
business logic config.